requirements.txt
click==8.0.4
colorama==0.4.5
dataclasses==0.8
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
importlib-metadata==4.8.3
itsdangerous==2.0.1
jinja2==3.0.0rc2
MarkupSafe==2.0.1
mysql-connector==2.2.9
numpy==1.19.5
pandas==0.25.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0